pyproject.toml
[project]
name = "ai-webpage-summarizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "litellm>=1.82.2",
    "mkdocs>=1.6.1",
    "mkdocs-awesome-pages-plugin>=2.10.1",
    "mkdocs-material>=9.7.5",
    "ollama>=0.6.1",
    "openai>=2.16.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
